internals: splitting up decls/defs for encapsulation related stuff
towards fixing #84
