+++ include/pfasst/encap/encapsulation.hpp
@@ -1,24 +1,22 @@
-/*
- * Host based encapsulated base sweeper.
- */
-
 #ifndef _PFASST_ENCAPSULATED_HPP_
 #define _PFASST_ENCAPSULATED_HPP_
 
-#include <vector>
 #include <memory>
-
-#include "../globals.hpp"
-#include "../interfaces.hpp"
-#include "../quadrature.hpp"
-
+#include <vector>
 using namespace std;
+
+#include <Eigen/Dense>
+
+template<typename scalar>
+using Matrix = Eigen::Matrix<scalar, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor>;
+
+#include "../interfaces.hpp"
+
 
 namespace pfasst
 {
   namespace encap
   {
-
     typedef enum EncapType { solution, function } EncapType;
 
     /**
@@ -34,52 +32,24 @@ namespace pfasst
     {
       public:
         //! @{
-        virtual ~Encapsulation()
-        {}
+        virtual ~Encapsulation();
         //! @}
 
         //! @{
         // required for time-parallel communications
-        virtual void post(ICommunicator* comm, int tag)
-        {
-          UNUSED(comm); UNUSED(tag);
-        }
-
-        virtual void send(ICommunicator* comm, int tag, bool blocking)
-        {
-          UNUSED(comm); UNUSED(tag); UNUSED(blocking);
-          throw NotImplementedYet("pfasst");
-        }
-
-        virtual void recv(ICommunicator* comm, int tag, bool blocking)
-        {
-          UNUSED(comm); UNUSED(tag); UNUSED(blocking);
-          throw NotImplementedYet("pfasst");
-        }
-
-        virtual void broadcast(ICommunicator* comm)
-        {
-          UNUSED(comm);
-          throw NotImplementedYet("pfasst");
-        }
+        virtual void post(ICommunicator* comm, int tag);
+        virtual void send(ICommunicator* comm, int tag, bool blocking);
+        virtual void recv(ICommunicator* comm, int tag, bool blocking);
+        virtual void broadcast(ICommunicator* comm);
         //! @}
 
         //! @{
         // required for host based encap helpers
-        virtual void zero()
-        {
-          throw NotImplementedYet("encap");
-        }
-        virtual void copy(shared_ptr<const Encapsulation<time>>)
-        {
-          throw NotImplementedYet("encap");
-        }
+        virtual void zero();
+        virtual void copy(shared_ptr<const Encapsulation<time>>);
         //! @}
 
-        virtual time norm0() const
-        {
-          throw NotImplementedYet("norm0");
-        }
+        virtual time norm0() const;
 
         //! @{
         /**
@@ -89,36 +59,18 @@ namespace pfasst
          * Here, \\(a\\) is a time point and \\(x\\) another data structure (usually of the
          * same type).
          */
-        virtual void saxpy(time a, shared_ptr<const Encapsulation<time>> x)
-        {
-          UNUSED(a); UNUSED(x);
-          throw NotImplementedYet("encap");
-        }
+        virtual void saxpy(time a, shared_ptr<const Encapsulation<time>> x);
 
         /**
          * defines matrix-vector multiplication for this data type.
          */
-        virtual void mat_apply(vector<shared_ptr<Encapsulation<time>>> dst, time a, Matrix<time> mat,
-                               vector<shared_ptr<Encapsulation<time>>> src, bool zero = true)
-        {
-          size_t ndst = dst.size();
-          size_t nsrc = src.size();
-
-          if (zero) {
-            for (auto elem : dst) { elem->zero(); }
-          }
-
-          for (size_t n = 0; n < ndst; n++) {
-            for (size_t m = 0; m < nsrc; m++) {
-              auto s = mat(n, m);
-              if (s != 0.0) {
-                dst[n]->saxpy(a*s, src[m]);
-              }
-            }
-          }
-        }
+        virtual void mat_apply(vector<shared_ptr<Encapsulation<time>>> dst,
+                               time a, Matrix<time> mat,
+                               vector<shared_ptr<Encapsulation<time>>> src,
+                               bool zero = true);
         //! @}
     };
+
 
     template<typename time = time_precision>
     class EncapFactory
@@ -126,8 +78,9 @@ namespace pfasst
       public:
         virtual shared_ptr<Encapsulation<time>> create(const EncapType) = 0;
     };
-
   }  // ::pfasst::encap
 } // ::pfasst
+
+#include "encapsulation_impl.hpp"
 
 #endif
