+++ include/pfasst/encap/imex_sweeper.hpp
@@ -1,20 +1,14 @@
-
 #ifndef _PFASST_ENCAP_IMEX_SWEEPER_HPP_
 #define _PFASST_ENCAP_IMEX_SWEEPER_HPP_
 
-#include <cstdlib>
-#include <cassert>
 #include <vector>
 #include <memory>
+using namespace std;
 
-#include "../globals.hpp"
-#include "../logging.hpp"
-#include "../quadrature.hpp"
 #include "encapsulation.hpp"
 #include "encap_sweeper.hpp"
 #include "vector.hpp"
 
-using namespace std;
 
 namespace pfasst
 {
@@ -25,13 +19,15 @@ namespace pfasst
     /**
      * Semi-implicit IMEX sweeper.
      *
-     * This IMEX sweeper is for ODEs of the form \\( \\dot{U} = F_{\\rm expl}(t,U) + F_{\\rm impl}(t, U)\\).  To reduce
-     * complexity and computational effort the non-stiff part is treated explicitly and the stiff part implicitly.
+     * This IMEX sweeper is for ODEs of the form
+     * \\( \\dot{U} = F_{\\rm expl}(t,U) + F_{\\rm impl}(t, U) \\).
+     * To reduce complexity and computational effort the non-stiff part is treated explicitly and 
+     * the stiff part implicitly.
      *
-     * This sweeper requires three interfaces to be implemented: two routines to evaluate the explicit \\( F_{\\rm expl}
-     * \\) and implicit \\( F_{\\rm impl} \\) pieces for a given state, and one routine that solves (perhaps with an
-     * external solver) the backward-Euler equation \\( U^{n+1} - \\Delta t F_{\\rm impl}(U^{n+1}) = RHS \\) for \\(
-     * U^{n+1} \\).
+     * This sweeper requires three interfaces to be implemented: two routines to evaluate the 
+     * explicit \\( F_{\\rm expl} \\) and implicit \\( F_{\\rm impl} \\) pieces for a given state, 
+     * and one routine that solves (perhaps with an external solver) the backward-Euler equation 
+     * \\( U^{n+1} - \\Delta t F_{\\rm impl}(U^{n+1}) = RHS \\) for \\( U^{n+1} \\).
      *
      * @tparam time precision type of the time dimension
      */
@@ -64,18 +60,11 @@ namespace pfasst
         /**
         * Set end state to \\( U_0 + \\int F_{expl} + F_{expl} \\).
         */
-        void integrate_end_state(time dt)
-        {
-          vector<shared_ptr<Encapsulation<time>>> dst = { this->end_state };
-          dst[0]->copy(this->start_state);
-          dst[0]->mat_apply(dst, dt, this->quadrature->get_b_mat(), this->fs_expl, false);
-          dst[0]->mat_apply(dst, dt, this->quadrature->get_b_mat(), this->fs_impl, false);
-        }
+        virtual void integrate_end_state(time dt);
 
       public:
         //! @{
         IMEXSweeper() = default;
-
         virtual ~IMEXSweeper() = default;
         //! @}
 
@@ -83,25 +72,7 @@ namespace pfasst
         /**
          * @copydoc ISweeper::setup(bool)
          */
-        virtual void setup(bool coarse) override
-        {
-          pfasst::encap::EncapSweeper<time>::setup(coarse);
-
-          auto const num_nodes = this->quadrature->get_num_nodes();
-          auto const num_s_integrals = this->quadrature->left_is_node() ? num_nodes - 1 : num_nodes;
-
-          for (size_t m = 0; m < num_nodes; m++) {
-            this->fs_expl.push_back(this->get_factory()->create(pfasst::encap::function));
-            this->fs_impl.push_back(this->get_factory()->create(pfasst::encap::function));
-          }
-          for (size_t m = 0; m < num_s_integrals; m++) {
-            this->s_integrals.push_back(this->get_factory()->create(pfasst::encap::solution));
-          }
-
-          if (! this->quadrature->left_is_node()) {
-            this->fs_expl_start = this->get_factory()->create(pfasst::encap::function);
-          }
-        }
+        virtual void setup(bool coarse) override;
 
         /**
          * Compute low-order provisional solution.
@@ -111,20 +82,7 @@ namespace pfasst
          * @param[in] initial if `true` the explicit and implicit part of the right hand side of the
          *     ODE get evaluated with the initial value
          */
-        virtual void predict(bool initial) override
-        {
-          if (this->quadrature->left_is_node()) {
-            this->predict_with_left(initial);
-          } else {
-            this->predict_without_left(initial);
-          }
-
-          if (this->quadrature->right_is_node()) {
-            this->end_state->copy(this->state.back());
-          } else {
-            this->integrate_end_state(this->get_controller()->get_time_step());
-          }
-        }
+        virtual void predict(bool initial) override;
 
         /**
          * Perform one SDC sweep/iteration.
@@ -132,56 +90,17 @@ namespace pfasst
          * This computes a high-order solution from the previous iteration's function values and
          * corrects it using forward/backward Euler steps across the nodes.
          */
-        virtual void sweep() override
-        {
-          if (this->quadrature->left_is_node()) {
-            this->sweep_with_left();
-          } else {
-            this->sweep_without_left();
-          }
-
-          if (this->quadrature->right_is_node()) {
-            this->end_state->copy(this->state.back());
-          } else {
-            this->integrate_end_state(this->get_controller()->get_time_step());
-          }
-        }
+        virtual void sweep() override;
 
         /**
          * Advance the end solution to start solution.
          */
-        virtual void advance() override
-        {
-          this->start_state->copy(this->end_state);
-          if (this->quadrature->left_is_node() && this->quadrature->right_is_node()) {
-            this->state[0]->copy(this->start_state);
-            this->fs_expl.front()->copy(this->fs_expl.back());
-            this->fs_impl.front()->copy(this->fs_impl.back());
-          }
-        }
+        virtual void advance() override;
 
         /**
          * @copybrief EncapSweeper::reevaluate()
          */
-        virtual void reevaluate(bool initial_only) override
-        {
-          time t0 = this->get_controller()->get_time();
-          time dt = this->get_controller()->get_time_step();
-          if (initial_only) {
-            if (this->quadrature->left_is_node()) {
-              this->f_expl_eval(this->fs_expl[0], this->state[0], t0);
-              this->f_impl_eval(this->fs_impl[0], this->state[0], t0);
-            } else {
-              throw NotImplementedYet("reevaluate");
-            }
-          } else {
-            for (size_t m = 0; m < this->quadrature->get_num_nodes(); m++) {
-              time t =  t0 + dt * this->quadrature->get_nodes()[m];
-              this->f_expl_eval(this->fs_expl[m], this->state[m], t);
-              this->f_impl_eval(this->fs_impl[m], this->state[m], t);
-            }
-          }
-        }
+        virtual void reevaluate(bool initial_only) override;
 
         /**
          * @copybrief EncapSweeper::integrate()
@@ -189,12 +108,7 @@ namespace pfasst
          * @param[in] dt width of time interval to integrate over
          * @param[in,out] dst integrated values; will get zeroed out beforehand
          */
-        virtual void integrate(time dt, vector<shared_ptr<Encapsulation<time>>> dst) const override
-        {
-          auto const q_mat = this->quadrature->get_q_mat();
-          dst[0]->mat_apply(dst, dt, q_mat, this->fs_expl, true);
-          dst[0]->mat_apply(dst, dt, q_mat, this->fs_impl, false);
-        }
+        virtual void integrate(time dt, vector<shared_ptr<Encapsulation<time>>> dst) const override;
 
         /**
          * @copybrief EncapSweeper::residual()
@@ -202,23 +116,7 @@ namespace pfasst
          * @param[in] dt width of time interval to integrate over
          * @param[in,out] dst residuals
          */
-        virtual void residual(time dt, vector<shared_ptr<Encapsulation<time>>> dst) const override
-        {
-          for (size_t m = 0; m < this->quadrature->get_num_nodes(); m++) {
-            dst[m]->copy(this->start_state);
-            dst[m]->saxpy(-1.0, this->state[m]);
-          }
-          if (this->fas_corrections.size() > 0) {
-            // XXX: build a matrix and use mat_apply to do this
-            for (size_t m = 0; m < this->quadrature->get_num_nodes(); m++) {
-              for (size_t n = 0; n <= m; n++) {
-                dst[m]->saxpy(1.0, this->fas_corrections[n]);
-              }
-            }
-          }
-          dst[0]->mat_apply(dst, dt, this->quadrature->get_q_mat(), this->fs_expl, false);
-          dst[0]->mat_apply(dst, dt, this->quadrature->get_q_mat(), this->fs_impl, false);
-        }
+        virtual void residual(time dt, vector<shared_ptr<Encapsulation<time>>> dst) const override;
         //! @}
 
         //! @{
@@ -234,11 +132,7 @@ namespace pfasst
          */
         virtual void f_expl_eval(shared_ptr<Encapsulation<time>> f_expl_encap,
                                  shared_ptr<Encapsulation<time>> u_encap,
-                                 time t)
-        {
-          UNUSED(f_expl_encap); UNUSED(u_encap); UNUSED(t);
-          throw NotImplementedYet("imex (f_expl_eval)");
-        }
+                                 time t);
 
         /**
          * Evaluate the implicit part of the ODE.
@@ -255,11 +149,7 @@ namespace pfasst
          */
         virtual void f_impl_eval(shared_ptr<Encapsulation<time>> f_impl_encap,
                                  shared_ptr<Encapsulation<time>> u_encap,
-                                 time t)
-        {
-          UNUSED(f_impl_encap); UNUSED(u_encap); UNUSED(t);
-          throw NotImplementedYet("imex (f_impl_eval)");
-        }
+                                 time t);
 
         /**
          * Solve \\( U - \\Delta t F_{\\rm impl}(U) = RHS \\) for \\( U \\).
@@ -280,164 +170,18 @@ namespace pfasst
         virtual void impl_solve(shared_ptr<Encapsulation<time>> f_encap,
                                 shared_ptr<Encapsulation<time>> u_encap,
                                 time t, time dt,
-                                shared_ptr<Encapsulation<time>> rhs_encap)
-        {
-          UNUSED(f_encap); UNUSED(u_encap); UNUSED(t); UNUSED(dt); UNUSED(rhs_encap);
-          throw NotImplementedYet("imex (impl_solve)");
-        }
+                                shared_ptr<Encapsulation<time>> rhs_encap);
         //! @}
 
-
       private:
-        void predict_with_left(bool initial)
-        {
-          time dt = this->get_controller()->get_time_step();
-          time t  = this->get_controller()->get_time();
-          CLOG(INFO, "Sweeper") << "predicting step " << this->get_controller()->get_step() + 1
-                                << " (t=" << t << ", dt=" << dt << ")";
-
-          if (initial) {
-            this->state[0]->copy(this->start_state);
-            this->f_expl_eval(this->fs_expl[0], this->state[0], t);
-            this->f_impl_eval(this->fs_impl[0], this->state[0], t);
-          }
-
-          shared_ptr<Encapsulation<time>> rhs = this->get_factory()->create(pfasst::encap::solution);
-
-          auto const nodes = this->quadrature->get_nodes();
-
-          // step across all nodes
-          for (size_t m = 0; m < nodes.size() - 1; ++m) {
-            time ds = dt * (nodes[m+1] - nodes[m]);
-            rhs->copy(this->state[m]);
-            rhs->saxpy(ds, this->fs_expl[m]);
-            this->impl_solve(this->fs_impl[m + 1], this->state[m + 1], t, ds, rhs);
-            this->f_expl_eval(this->fs_expl[m + 1], this->state[m + 1], t + ds);
-            t += ds;
-          }
-        }
-
-        void predict_without_left(bool initial)
-        {
-          UNUSED(initial);
-          time dt = this->get_controller()->get_time_step();
-          time t  = this->get_controller()->get_time();
-          CLOG(INFO, "Sweeper") << "predicting step " << this->get_controller()->get_step() + 1
-                                << " (t=" << t << ", dt=" << dt << ")";
-          time ds;
-
-          shared_ptr<Encapsulation<time>> rhs = this->get_factory()->create(pfasst::encap::solution);
-
-          auto const nodes = this->quadrature->get_nodes();
-
-          // step to first node
-          ds = dt * nodes[0];
-          this->f_expl_eval(this->fs_expl_start, this->start_state, t);
-          rhs->copy(this->start_state);
-          rhs->saxpy(ds, this->fs_expl_start);
-          this->impl_solve(this->fs_impl[0], this->state[0], t, ds, rhs);
-          this->f_expl_eval(this->fs_expl[0], this->state[0], t + ds);
-
-          // step across all nodes
-          for (size_t m = 0; m < nodes.size() - 1; ++m) {
-            ds = dt * (nodes[m+1] - nodes[m]);
-            rhs->copy(this->state[m]);
-            rhs->saxpy(ds, this->fs_expl[m]);
-            this->impl_solve(this->fs_impl[m+1], this->state[m+1], t, ds, rhs);
-            this->f_expl_eval(this->fs_expl[m+1], this->state[m+1], t + ds);
-            t += ds;
-          }
-        }
-
-        void sweep_with_left()
-        {
-          auto const nodes = this->quadrature->get_nodes();
-          auto const dt    = this->get_controller()->get_time_step();
-          auto const s_mat = this->quadrature->get_s_mat().block(1, 0, nodes.size()-1, nodes.size());
-          CLOG(INFO, "Sweeper") << "sweeping on step " << this->get_controller()->get_step() + 1
-                                << " in iteration " << this->get_controller()->get_iteration() << " (dt=" << dt << ")";
-          time ds;
-
-          this->s_integrals[0]->mat_apply(this->s_integrals, dt, s_mat, this->fs_expl, true);
-          this->s_integrals[0]->mat_apply(this->s_integrals, dt, s_mat, this->fs_impl, false);
-          for (size_t m = 0; m < nodes.size() - 1; ++m) {
-            ds = dt * (nodes[m+1] - nodes[m]);
-            this->s_integrals[m]->saxpy(-ds, this->fs_expl[m]);
-            this->s_integrals[m]->saxpy(-ds, this->fs_impl[m+1]);
-          }
-          if (this->fas_corrections.size() > 0) {
-            for (size_t m = 0; m < this->s_integrals.size(); m++) {
-              this->s_integrals[m]->saxpy(1.0, this->fas_corrections[m+1]);
-            }
-          }
-
-          shared_ptr<Encapsulation<time>> rhs = this->get_factory()->create(pfasst::encap::solution);
-
-          // step across all nodes
-          auto t = this->get_controller()->get_time();
-          for (size_t m = 0; m < nodes.size() - 1; ++m) {
-            ds = dt * (nodes[m+1] - nodes[m]);
-            rhs->copy(this->state[m]);
-            rhs->saxpy(ds, this->fs_expl[m]);
-            rhs->saxpy(1.0, this->s_integrals[m]);
-            this->impl_solve(this->fs_impl[m + 1], this->state[m + 1], t, ds, rhs);
-            this->f_expl_eval(this->fs_expl[m + 1], this->state[m + 1], t + ds);
-            t += ds;
-          }
-        }
-
-        void sweep_without_left()
-        {
-          auto const nodes = this->quadrature->get_nodes();
-          auto const dt    = this->get_controller()->get_time_step();
-          auto const s_mat = this->quadrature->get_s_mat();
-          CLOG(INFO, "Sweeper") << "sweeping on step " << this->get_controller()->get_step() + 1
-                                << " in iteration " << this->get_controller()->get_iteration() << " (dt=" << dt << ")";
-          time ds;
-
-          this->s_integrals[0]->mat_apply(this->s_integrals, dt, s_mat, this->fs_expl, true);
-          this->s_integrals[0]->mat_apply(this->s_integrals, dt, s_mat, this->fs_impl, false);
-          ds = dt * nodes[0];
-          this->s_integrals[0]->saxpy(-ds, this->fs_expl_start);
-          this->s_integrals[0]->saxpy(-ds, this->fs_impl[0]);
-          for (size_t m = 0; m < nodes.size() - 1; ++m) {
-            ds = dt * (nodes[m+1] - nodes[m]);
-            this->s_integrals[m+1]->saxpy(-ds, this->fs_expl[m]);
-            this->s_integrals[m+1]->saxpy(-ds, this->fs_impl[m+1]);
-          }
-          if (this->fas_corrections.size() > 0) {
-            for (size_t m = 0; m < this->s_integrals.size(); m++) {
-              this->s_integrals[m]->saxpy(1.0, this->fas_corrections[m]);
-            }
-          }
-
-          shared_ptr<Encapsulation<time>> rhs = this->get_factory()->create(pfasst::encap::solution);
-
-          // step to first node
-          auto t = this->get_controller()->get_time();
-          ds = dt * nodes[0];
-          this->f_expl_eval(this->fs_expl_start, this->start_state, t);
-          rhs->copy(this->start_state);
-          rhs->saxpy(ds, this->fs_expl_start);
-          rhs->saxpy(1.0, this->s_integrals[0]);
-          this->impl_solve(this->fs_impl[0], this->state[0], t, ds, rhs);
-          this->f_expl_eval(this->fs_expl[0], this->state[0], t + ds);
-
-          // step across all nodes
-          for (size_t m = 0; m < nodes.size() - 1; ++m) {
-            ds = dt * (nodes[m+1] - nodes[m]);
-            rhs->copy(this->state[m]);
-            rhs->saxpy(ds, this->fs_expl[m]);
-            rhs->saxpy(1.0, this->s_integrals[m+1]);
-            this->impl_solve(this->fs_impl[m+1], this->state[m+1], t, ds, rhs);
-            this->f_expl_eval(this->fs_expl[m+1], this->state[m+1], t + ds);
-            t += ds;
-          }
-        }
-
+        virtual void predict_with_left(bool initial);
+        virtual void predict_without_left(bool initial);
+        virtual void sweep_with_left();
+        virtual void sweep_without_left();
     };
-
   }  // ::pfasst::encap
 }  // ::pfasst
+
+#include "imex_sweeper_impl.hpp"
 
 #endif
